Make lag time calculation robust against recent missing raw measurements
The code for setting the zero time based on the requested lag was doing so in a
rather haphazard way by going to the "lag"th-last raw data point. This can
unintentionally increase the lag if recent data points are missing. The new code
instead tries to find a raw measurement with a timestamp that is within a
prescribed tolerance of the desired lag, and if there is none, it directly
calculates the correct timestamp for the requested lag.

The code now also checks that there is at least one raw measurement that is older
than the timestamp with the requested lag and errors out otherwise.